+++ src/Dto/Cronjob/ConfigSettingsDto.php
@@ -15,9 +15,9 @@ class ConfigSettingsDto implements SettingInterface
 {
     use SettingsTrait;
 
-    private bool $log;
-    private string $destinationFile;
-    private bool $heartbeat;
+    private bool $log = false;
+    private string $destinationFile = '';
+    private bool $heartbeat = false;
     private ?string $user = null;
 
     /** @param array<string, mixed> $settings */