+++ src/Dto/Trait/SettingsTrait.php
@@ -50,7 +50,7 @@ trait SettingsTrait
 
     private function toCamelCase(string $input): string
     {
-        $camelCaseString = \str_replace(' ', '', \ucwords(\str_replace('_', ' ', $input)));
+        $camelCaseString = \str_replace(' ', '', \ucwords(\str_replace(['_', '-'], ' ', $input)));
 
         return \lcfirst($camelCaseString);
     }